+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- `toolbox.visualisation.Plotcontext` now can plot vertical apertures too.
 - `toolbox.realign_mad_apertures()` now realigns vertical rectangular apertures too.
 
 ## [0.1.0] - 16-09-2026